third_party/blink/web_tests/http/tests/misc/resource-timing-iframe-restored-from-history.html

<!DOCTYPE html>
<script src="/js-test-resources/js-test.js"></script>
<script src="/resources/prevent-bfcache.js"></script>
<script>
description('Tests that iframe restored from history without explicit navigation does not report resource timing.');
window.jsTestIsAsync = true;

async function runTest() {
    await preventBFCache();
    if (!sessionStorage.didNav) {
        sessionStorage.didNav = true;
        // Navigate a timeout to make sure we generate a history entry that we can go back to.
        setTimeout(function() {
            location.href = 'resources/alert-then-back.html';
        }, 0);
    } else {
        delete sessionStorage.didNav;
        resources = performance.getEntriesByType('resource');
        shouldBe('resources.length', '2');
        shouldBeEqualToString('resources[0].name', 'http://127.0.0.1:8000/js-test-resources/js-test.js');
        shouldBeEqualToString('resources[1].name', 'http://127.0.0.1:8000/resources/prevent-bfcache.js');
        if (window.testRunner)
            finishJSTest();
    }
}

if (!sessionStorage.didNav) {
  window.onload = function() {
    document.querySelector("iframe").src = "resources/frame-initial-url.html";
  }
}

</script>

<p>Back-forward cache keeps and restores the resource timing of the iframe on
back navigation. Back-forward cache needs to be disabled to run this test.
</p>
<iframe></iframe>
